+++ examples/ai-core/src/e2e/google-vertex.test.ts
@@ -101,7 +101,7 @@ const createModelsForRuntime = (
   ],
   imageModels: [
     createImageModel(vertex, 'imagen-3.0-fast-generate-001', [imageTest]),
-    createImageModel(vertex, 'imagen-3.0-generate-001', [imageTest]),
+    createImageModel(vertex, 'imagen-3.0-generate-002', [imageTest]),
   ],
 });