+++ packages/react/src/use-chat.ts
@@ -52,6 +52,12 @@ export type UseChatHelpers = {
    * Abort the current request immediately, keep the generated tokens if any.
    */
   stop: () => void;
+
+  /**
+   * Resume an ongoing chat generation stream. This does not resume an aborted generation.
+   */
+  experimental_resume: () => void;
+
   /**
    * Update the `messages` state locally. This is useful when you want to
    * edit the messages on the client, and then trigger the `reload` method
@@ -236,7 +242,10 @@ By default, it's set to 1, which means that only a single LLM call is made.
   }, [credentials, headers, body]);
 
   const triggerRequest = useCallback(
-    async (chatRequest: ChatRequest) => {
+    async (
+      chatRequest: ChatRequest,
+      requestType: 'generate' | 'resume' = 'generate',
+    ) => {
       mutateStatus('submitted');
       setError(undefined);
 
@@ -339,6 +348,7 @@ By default, it's set to 1, which means that only a single LLM call is made.
           generateId,
           fetch,
           lastMessage: chatMessages[chatMessages.length - 1],
+          requestType,
         });
 
         abortControllerRef.current = null;
@@ -407,7 +417,7 @@ By default, it's set to 1, which means that only a single LLM call is made.
         data,
         headers,
         body,
-        experimental_attachments,
+        experimental_attachments = message.experimental_attachments,
       }: ChatRequestOptions = {},
     ) => {
       const attachmentsForRequest = await prepareAttachmentsForRequest(
@@ -455,6 +465,12 @@ By default, it's set to 1, which means that only a single LLM call is made.
       abortControllerRef.current = null;
     }
   }, []);
+
+  const experimental_resume = useCallback(async () => {
+    const messages = messagesRef.current;
+
+    triggerRequest({ messages }, 'resume');
+  }, [triggerRequest]);
 
   const setMessages = useCallback(
     (messages: Message[] | ((messages: Message[]) => Message[])) => {
@@ -581,6 +597,7 @@ By default, it's set to 1, which means that only a single LLM call is made.
     append,
     reload,
     stop,
+    experimental_resume,
     input,
     setInput,
     handleInputChange,