+++ packages/react/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ai-sdk/react
 
+## 1.2.12
+
+### Patch Changes
+
+- Updated dependencies [d87b9d1]
+  - @ai-sdk/provider-utils@2.2.8
+  - @ai-sdk/ui-utils@1.2.11
+
+## 1.2.11
+
+### Patch Changes
+
+- 6c59ae7: feat (ui/react): support resuming an ongoing stream
+- Updated dependencies [6c59ae7]
+  - @ai-sdk/ui-utils@1.2.10
+
+## 1.2.10
+
+### Patch Changes
+
+- 77b2097: fix(append): support Message.experimental_attachments alongside ChatRequestOptions
+- Updated dependencies [62181ef]
+  - @ai-sdk/ui-utils@1.2.9
+
 ## 1.2.9
 
 ### Patch Changes