+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/svelte
 
+## 2.1.12
+
+### Patch Changes
+
+- Updated dependencies [d87b9d1]
+  - @ai-sdk/provider-utils@2.2.8
+  - @ai-sdk/ui-utils@1.2.11
+
+## 2.1.11
+
+### Patch Changes
+
+- Updated dependencies [6c59ae7]
+  - @ai-sdk/ui-utils@1.2.10
+
+## 2.1.10
+
+### Patch Changes
+
+- Updated dependencies [62181ef]
+  - @ai-sdk/ui-utils@1.2.9
+
 ## 2.1.9
 
 ### Patch Changes